+++ benchmark/dynamic-array/performance.js
@@ -1,47 +1,52 @@
 var DynamicArray = require('../../dynamic-array.js');
 var HashedArrayTree = require('../../hashed-array-tree.js');
 
-var SIZE = 20000000;
+var SIZE = 50000000,
+    MAX_VALUE = 256,
+    ARRAY_TYPE = Uint8Array;
 
 console.time('Standard JavaScript Array (holding uint8) #push');
 var array = [];
 
 for (var i = 0; i < SIZE; i++)
-  array.push(i % 256);
+  array.push(i % MAX_VALUE);
 
 console.timeEnd('Standard JavaScript Array (holding uint8) #push');
 
 console.time('DynamicUint8Array #push');
-var dynamicArray = new DynamicArray(Uint8Array, {
-  initialSize: 10
-});
+var dynamicArray = new DynamicArray(ARRAY_TYPE);
 
 for (var i = 0; i < SIZE; i++)
-  dynamicArray.push(i % 256);
+  dynamicArray.push(i % MAX_VALUE);
 console.timeEnd('DynamicUint8Array #push');
 
 console.time('HashedArrayTree uint8 #push');
-var hashedArrayTree = new HashedArrayTree(Uint8Array);
+var hashedArrayTree = new HashedArrayTree(ARRAY_TYPE);
 
 for (var i = 0; i < SIZE; i++)
-  hashedArrayTree.push(i % 256);
+  hashedArrayTree.push(i % MAX_VALUE);
 console.timeEnd('HashedArrayTree uint8 #push');
 
 console.log();
 
 console.time('Standard JavaScript Array (holding uint8) #set');
 for (var i = 0; i < SIZE; i++)
-  array[i] = i % 256;
+  array[i] = i % MAX_VALUE;
 console.timeEnd('Standard JavaScript Array (holding uint8) #set');
 
 console.time('DynamicUint8Array #set');
 for (var i = 0; i < SIZE; i++)
-  dynamicArray.set(i, i % 256);
+  dynamicArray.set(i, i % MAX_VALUE);
 console.timeEnd('DynamicUint8Array #set');
+
+console.time('DynamicUint8Array #set(fast)');
+for (var i = 0; i < SIZE; i++)
+  dynamicArray.array[i] = i % MAX_VALUE;
+console.timeEnd('DynamicUint8Array #set(fast)');
 
 console.time('HashedArrayTree uint8 #set');
 for (var i = 0; i < SIZE; i++)
-  hashedArrayTree.set(i, i % 256);
+  hashedArrayTree.set(i, i % MAX_VALUE);
 console.timeEnd('HashedArrayTree uint8 #set');
 
 console.log();
@@ -56,6 +61,11 @@ console.time('DynamicUint8Array #get');
 for (var i = 0; i < SIZE; i++)
   v = dynamicArray.get(i);
 console.timeEnd('DynamicUint8Array #get');
+
+console.time('DynamicUint8Array #get(fast)');
+for (var i = 0; i < SIZE; i++)
+  v = dynamicArray.array[i];
+console.timeEnd('DynamicUint8Array #get(fast)');
 
 console.time('HashedArrayTree uint8 #get');
 for (var i = 0; i < SIZE; i++)