+++ test/dynamic-array.js
@@ -22,7 +22,7 @@ describe('DynamicArray', function() {
   });
 
   it('should be possible to set and get values.', function() {
-    var array = new DynamicArray(Uint8Array, 5);
+    var array = new DynamicArray(Uint8Array, {initialLength: 3});
 
     array.set(2, 24);
 
@@ -37,26 +37,34 @@ describe('DynamicArray', function() {
     assert.strictEqual(array.get(2), undefined);
   });
 
-  it('setting an out-of-bound value should grow the array.', function() {
-    var array = new DynamicArray(Uint8Array, {
-      initialCapacity: 2,
-      policy: function(capacity) {
-        return capacity + 2;
-      }
-    });
+  // it('setting an out-of-bound index should grow the array.', function() {
+  //   var array = new DynamicArray(Uint8Array, {
+  //     initialCapacity: 2,
+  //     policy: function(capacity) {
+  //       return capacity + 2;
+  //     }
+  //   });
 
-    array.set(3, 4);
+  //   array.set(3, 4);
 
-    assert.strictEqual(array.length, 4);
-    assert.strictEqual(array.get(3), 4);
-    assert.deepEqual(Array.from(array.array), [0, 0, 0, 4]);
+  //   assert.strictEqual(array.length, 4);
+  //   assert.strictEqual(array.get(3), 4);
+  //   assert.deepEqual(Array.from(array.array), [0, 0, 0, 4]);
 
-    array.set(6, 22);
+  //   array.set(6, 22);
 
-    assert.strictEqual(array.length, 7);
-    assert.strictEqual(array.get(6), 22);
-    assert.strictEqual(array.get(2), 0);
-    assert.deepEqual(Array.from(array.array), [0, 0, 0, 4, 0, 0, 22, 0]);
+  //   assert.strictEqual(array.length, 7);
+  //   assert.strictEqual(array.get(6), 22);
+  //   assert.strictEqual(array.get(2), 0);
+  //   assert.deepEqual(Array.from(array.array), [0, 0, 0, 4, 0, 0, 22, 0]);
+  // });
+
+  it('setting an out-of-bound index should throw.', function() {
+    var array = new DynamicArray(Uint8Array, 4);
+
+    assert.throws(function() {
+      array.set(56, 4);
+    }, /bounds/);
   });
 
   it('should be possible to push values.', function() {
@@ -88,6 +96,59 @@ describe('DynamicArray', function() {
 
     assert.strictEqual(array.get(1), 35);
     assert.strictEqual(array.length, 2);
+  });
+
+  it('should be possible to reallocate.', function() {
+    var array = new DynamicArray(Uint8Array, 10);
+
+    array.push(1);
+    array.push(2);
+    array.push(3);
+
+    array.reallocate(20);
+
+    assert.strictEqual(array.capacity, 20);
+    assert.strictEqual(array.length, 3);
+
+    array.reallocate(2);
+
+    assert.strictEqual(array.capacity, 2);
+    assert.strictEqual(array.length, 2);
+  });
+
+  it('should be possible to grow the array.', function() {
+    var array = new DynamicArray(Uint8Array, {
+      initialCapacity: 2,
+      policy: function(capacity) {
+        return capacity + 2;
+      }
+    });
+
+    array.grow(5);
+
+    assert.strictEqual(array.capacity, 6);
+
+    array.grow(2);
+
+    assert.strictEqual(array.capacity, 6);
+
+    array.grow();
+
+    assert.strictEqual(array.capacity, 8);
+  });
+
+  it('should be possible to resize the array.', function() {
+    var array = new DynamicArray(Uint8Array, {initialLength: 23});
+
+    array.resize(20);
+
+    assert.strictEqual(array.capacity, 23);
+    assert.strictEqual(array.length, 20);
+
+    array.resize(30);
+
+    assert.strictEqual(array.capacity, 30);
+    assert.strictEqual(array.length, 30);
   });
 
   it('should throw if the policy returns an irrelevant size.', function() {
@@ -122,7 +183,7 @@ describe('DynamicArray', function() {
   });
 
   it('should be possible to use the subclasses.', function() {
-    var array = new DynamicArray.DynamicFloat64Array(4);
+    var array = new DynamicArray.DynamicFloat64Array({initialLength: 3});
 
     array.set(2, 24);